+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Load V2 configuration for all services in parallel [PR #272](https://github.com/3scale/apicast/pull/272)
 - Configuration is using LRU cache [PR #274](https://github.com/3scale/apicast/pull/274)
 - Management API not opened by default [PR #276](https://github.com/3scale/apicast/pull/276)
+- Management API returns ready status with no services [PR #]()
 
 ### Added